Simplify Dockerfile and streamline build process
This commit simplifies the Dockerfile and streamlines the build process. The separate build and install stages for the Python wheel have been removed. Instead, the poetry installation and package installation are now done directly in the 'prod' stage. This reduces the complexity of the Dockerfile and makes the build process faster and more efficient.

The entrypoint script path has also been updated to reflect its location within the Docker context. This change makes the Dockerfile more portable and easier to use in different environments.